Move theme classes to theme lib
