+++ android/lib/theme/src/main/kotlin/net/mullvad/mullvadvpn/lib/theme/Type.kt
@@ -1,4 +1,4 @@
-package net.mullvad.mullvadvpn.compose.theme
+package net.mullvad.mullvadvpn.lib.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
